[BLACKTIE-320] Add priority test for  externally managed queues
git-svn-id: http://anonsvn.jboss.org/repos/blacktie/trunk@2307 aa7494b7-e75f-0410-880c-f61b6724bf28

+++ atmibroker-xatmi/src/test/cpp/TestTPCall.cxx
@@ -298,7 +298,7 @@ void TestTPCall::test_tpcall_with_TPNOCHANGE() {
 
 	char* toTestS = (char*) malloc(110);
 	sprintf(toTestS, "%ld", toTest);
-	BT_ASSERT_MESSAGE(toTestS, toTest >= sizeof(ACCT_INFO));
+	BT_ASSERT_MESSAGE(toTestS, toTest >= (int) sizeof(ACCT_INFO));
 	free(toTestS);
 
 	free(type);
